backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "cryptography>=46.0.3",
    "fastapi>=0.128.0",
    "gitpython>=3.1.46",
    "greenlet>=3.3.0",
    "httpx>=0.28.1",
    "itsdangerous>=2.2.0",
    "ptyprocess>=0.7.0",
    "pydantic-settings>=2.12.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.21",
    "sqlalchemy>=2.0.45",
    "uvicorn[standard]>=0.40.0",
    "websockets>=15.0.1",
]
